chore: Renaming a few files to stick to prior pluralization convetions
