+++ client/src/store/rootReducer.js
@@ -1,5 +1,5 @@
 import { combineReducers } from "redux";
-import propertiesListReducer from "./property/propertySlice";
+import propertiesListReducer from "./property/propertiesSlice";
 
 export default combineReducers({
 	propertiesList: propertiesListReducer,
